requirements.txt
flask==3.1.1 
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-dotenv==1.1.0
PyMySQL==1.1.0
cryptography==42.0.0
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2
Flask-WTF==1.2.1
WTForms==3.1.1
email_validator==2.1.1